requirements.txt
django>=6.0,<7.0
djangorestframework>=3.17,<4.0
pandas>=3.0,<4.0
requests>=2.33,<3.0
drf-spectacular>=0.29,<1.0
pytest>=9.0,<10.0
